# Make stage 2 structural iteration stackless