--- a/src/arm64/dom_parser_implementation.cpp
+++ b/src/arm64/dom_parser_implementation.cpp
@@ -146,14 +146,13 @@ WARN_UNUSED bool implementation::validate_utf8(const char *buf, size_t len) cons
 
 WARN_UNUSED error_code dom_parser_implementation::stage2(dom::document &_doc) noexcept {
   doc = &_doc;
-  stage2::tape_builder builder(*doc);
-  return stage2::structural_parser::parse<false>(*this, builder);
+  SIMDJSON_ASSUME(next_structural_index == 0);
+  return stage2::tape_builder::parse<false>(*this, 0);
 }
 
 WARN_UNUSED error_code dom_parser_implementation::stage2_next(dom::document &_doc) noexcept {
   doc = &_doc;
-  stage2::tape_builder builder(_doc);
-  return stage2::structural_parser::parse<true>(*this, builder);
+  return stage2::tape_builder::parse<true>(*this, next_structural_index);
 }
 
 WARN_UNUSED error_code dom_parser_implementation::parse(const uint8_t *_buf, size_t _len, dom::document &_doc) noexcept {