--- a/src/arm64/dom_parser_implementation.cpp
+++ b/src/arm64/dom_parser_implementation.cpp
@@ -111,7 +111,6 @@ really_inline simd8<bool> must_be_2_3_continuation(const simd8<uint8_t> prev2, c
 
 #include "arm64/stringparsing.h"
 #include "arm64/numberparsing.h"
-#include "generic/stage2/structural_parser.h"
 #include "generic/stage2/tape_builder.h"
 
 //