--- a/src/arm64/dom_parser_implementation.cpp
+++ b/src/arm64/dom_parser_implementation.cpp
@@ -146,12 +146,12 @@ WARN_UNUSED bool implementation::validate_utf8(const char *buf, size_t len) cons
 WARN_UNUSED error_code dom_parser_implementation::stage2(dom::document &_doc) noexcept {
   doc = &_doc;
   SIMDJSON_ASSUME(next_structural_index == 0);
-  return stage2::tape_builder::parse<false>(*this, 0);
+  return stage2::tape_builder::parse<false>(*this);
 }
 
 WARN_UNUSED error_code dom_parser_implementation::stage2_next(dom::document &_doc) noexcept {
   doc = &_doc;
-  return stage2::tape_builder::parse<true>(*this, next_structural_index);
+  return stage2::tape_builder::parse<true>(*this);
 }
 
 WARN_UNUSED error_code dom_parser_implementation::parse(const uint8_t *_buf, size_t _len, dom::document &_doc) noexcept {