… src/generic/stage2/structural_parser.h
… src/haswell/dom_parser_implementation.cpp
… src/westmere/dom_parser_implementation.cpp