This is an experiment with making stage 2 work *without* having to keep a stack telling it what is a container and what is an array. The key change is a [new state called `generic_next`](https://github.com/simdjson/simdjson/blob/f1f9927f2dd29f139b4e151210b1b3764bddbb02/src/generic/stage2/structural_parser.h#L185-L228), which we switch into when we hit a `}` or `]`. This state reads the next few tokens to figure out what state it should be in: if there is a `, "key" :` we go back to `object_value` state; if there is a comma followed by anything else, we treat it as an array value. We call `try_resume_array` / `try_resume_object` on the builder to give it a chance to fault us immediately if it wants to.

You'll notice the tape_builder still handles depth and is_array. I think at least some of that can be optimized away without losing anything. We shall see.

## Interface

This makes the SAX interface nicer as well (which is actually what I believe causes the performance impacts), allowing for object field optimizations in particular (you often want to know what key you have before interpreting the value).

Single-value interface: the basic value callbacks. Each of these has a _field() variant that is also passed the key (we pass key and value together), and a _document() variant that indicates it was found at the root and comprises the entire document.
* primitive(value) / primitive_field(key, value) / primitive_document(value) - String, number, or atom.
* empty_array / empty_array_field(key) / empty_array_document - empty array or object.
* empty_object / empty_object_field(key) empty_object_document - empty array or object.
* error - Error detected.

Container interface: these methods help open and out or resume objects and arrays and give the opportunity to validate balance.

* start_array / start_array_field(key) / start_array_document - Start of non-empty array.
* start_object / start_object_field(key) / start_object_document - Start of non-empty object.
* end_array / end_array_document - end of a non-empty array.
* end_object / end_object_document - end of a non-empty object.
* try_end_array / try_end_object / try_end_array_document / try_end_object_document - End of non-empty array or object that may not be balanced.
* try_resume_array / try_resume_object - signal that we've found more array elements / objects, but the array/object may not be balanced.
* try_resume_array(string_value) / try_end_object(string_value) - resume or end an array, and we found a string value next.

## Performance

After just refactoring to the stackless iterator (but keeping the stack in the tape_builder instead), it's still a bit slower; but we're not taking advantage of the stacklessness, so that's not surprising. The fact that it's this close makes me think something is there (and I have reason to believe a lot of the issue has more to do with the refactoring of the interface than the stackless state).

skylake-x gcc10.2:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| mesh.pretty                      |  24646 | 173.4 | 174.0 |    0% | 581.9 | 588.7 |   -1% |
| citm_catalog                     |  26987 |  79.1 |  80.4 |   -1% | 293.5 | 308.0 |   -4% |
| twitter                          |   9867 |  90.4 |  91.8 |   -1% | 287.3 | 298.7 |   -3% |
| mesh                             |  11306 | 289.2 | 294.4 |   -1% | 994.3 | 1010.9 |   -1% |
| gsoc-2018                        |  51997 |  72.6 |  74.1 |   -2% | 164.9 | 170.8 |   -3% |
| github_events                    |   1017 |  77.0 |  79.1 |   -2% | 262.9 | 274.4 |   -4% |
| marine_ik                        |  46616 | 275.8 | 285.1 |   -3% | 907.0 | 931.6 |   -2% |
| instruments                      |   3442 | 103.2 | 107.0 |   -3% | 376.2 | 393.1 |   -4% |
| random                           |   7976 | 137.8 | 143.0 |   -3% | 493.3 | 516.9 |   -4% |
| apache_builds                    |   1988 |  82.8 |  86.0 |   -3% | 303.9 | 321.5 |   -5% |
| update-center                    |   8330 | 110.5 | 115.1 |   -4% | 334.6 | 352.1 |   -5% |
| twitterescaped                   |   8787 | 183.7 | 193.3 |   -5% | 494.4 | 514.0 |   -3% |
| numbers                          |   2345 | 237.8 | 252.3 |   -6% | 803.6 | 799.3 |    0% |
| canada                           |  35172 | 269.2 | 287.2 |   -6% | 957.3 | 979.6 |   -2% |