@jkeiser 2020-08-06
I'm still pretty bullish on this, but I don't think it's going to help with interfaces we have in the short term. You have to know what is and isn't an array if you want to validate. If we just let the tape_builder pass that information back to the parser when a container is ended, that should keep the structural_iterator itself stackless and let the visitor do the stack however they want (some people will actually have parent pointers in their containers naturally and won't need a stack).