@jkeiser 2020-08-06
I'll run another pass on this without the stackless SAX; I did like the stack separation and the visitor interface.