@jkeiser 2020-08-13
@lemire I'll close this one pretty soon, but it's an interesting design to keep in the back of your head when you do get a moment to look :)