@lemire 2020-08-14
@jkeiser I have reviewed it. It is interesting.

Don't delete the branch.